+++ src/versioned-entity/event-processor.ts
@@ -1,8 +1,15 @@
 import type { Type } from '@nestjs/common';
+import { plainToInstance } from 'class-transformer';
+import {
+  EntityNotFoundError,
+  OldEntityVersionError,
+  UnsupportedEntityOperationError,
+} from '../errors/index.js';
 import {
   Transaction,
   TransactionRunner,
   type ReadOnlyStateTransaction,
+  type ReadOnlyTransactionOption,
   type TransactionOption,
 } from '../transaction/index.js';
 import type { KeyOfType } from '../typing/index.js';
@@ -12,11 +19,71 @@ import type { KeyOfType } from '../typing/index.js';
  */
 export type VersionedProjectionOptions<P extends object> = {
   /**
-   * The existing state before processing the event.
-   * This is only passed if {@link VersionedEventProcessor.stateKeyForEvent} is implemented and returns a key, and when
-   * the state already exists.
+   * The existing state before processing the event, forwarded from {@link VersionedEventProcessor.processEvent}.
+   * See {@link VersionedEventProcessingOptions.existingState}.
    */
-  state?: P;
+  existingState?: P | null;
+};
+
+/**
+ * A function that fetches the version of a projection.
+ */
+export type ProjectionVersionFetcher<P extends object> = (
+  projection: P,
+) => Date | undefined | null;
+
+/**
+ * A context returned by {@link VersionedEventProcessor.project} that can contain additional information to take into
+ * account for processing.
+ */
+export type ProjectionContext<P extends object, K extends keyof P = keyof P> = {
+  /**
+   * The existing state of the projection in the database.
+   * This should be passed if the state was fetched from the database as part of the projection.
+   * Passing `null` means that the state does not currently exist in the database.
+   */
+  existingState?: P | null;
+
+  /**
+   * A initial state used when the projection function returns a partial object.
+   * If provided, either the null projection or the existing state will be used to fill in the missing properties.
+   */
+  defaultProjection?: Omit<P, K>;
+
+  /**
+   * A function that returns the version property of the projection, used to compare the projection's version with the
+   * existing state.
+   */
+  projectionVersionProperty?: ProjectionVersionFetcher<P>;
+};
+
+/**
+ * The return type of a {@link VersionedEventProcessor.project} method that provides a {@link ProjectionContext}.
+ */
+export type ProjectionWithContext<
+  P extends object,
+  K extends keyof P = keyof P,
+> = [Pick<P, K> & Partial<P>, ProjectionContext<P, K>];
+
+/**
+ * Options for processing a versioned event.
+ */
+export type VersionedEventProcessingOptions<
+  RWT extends Transaction,
+  P extends object,
+> = TransactionOption<RWT> & {
+  /**
+   * The existing state of the projection in the database.
+   * If provided, this is used instead of fetching the state from the database.
+   * Passing `null` means that the state does not currently exist in the database.
+   */
+  existingState?: P | null;
+
+  /**
+   * If `true`, the processor will reprocess the event even if the existing state has the same version as the event.
+   * This can be useful for backfilling cases.
+   */
+  reprocessEqualVersion?: boolean;
 };
 
 /**
@@ -34,31 +101,27 @@ export abstract class VersionedEventProcessor<
   R extends TransactionRunner<RWT, ROT> = TransactionRunner<RWT, ROT>,
 > {
   /**
+   * The function to use to fetch the version of a projection.
+   */
+  readonly projectionVersionProperty: ProjectionVersionFetcher<P>;
+
+  /**
    * Creates a new {@link VersionedEventProcessor}.
    *
    * @param projectionType The class of the state to build, which should be a type accepted by the state transaction.
    *   This is usually equal to the data in the event, but it can be different.
    * @param runner The {@link TransactionRunner} used to create transactions.
-   * @param projectionVersionProperty The name of the property that should be used to compare the state and the
-   *   projection's versions.
+   * @param projectionVersionProperty The property or function to use to fetch the version of the projection or state.
    */
   constructor(
     readonly projectionType: Type<P>,
     readonly runner: R,
-    readonly projectionVersionProperty: KeyOfType<P, Date>,
-  ) {}
-
-  /**
-   * A method that returns the key used to fetch the existing state for the given event.
-   * This only needs to be implemented if the existing state should be passed to
-   * {@link VersionedEventProcessor.project}.
-   *
-   * @param event The event for which the key should be built.
-   * @returns The partial projection, containing the properties defining the key for the projection / state.
-   */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  protected stateKeyForEvent(event: E): Partial<P> | null {
-    return null;
+    projectionVersionProperty: KeyOfType<P, Date> | ProjectionVersionFetcher<P>,
+  ) {
+    this.projectionVersionProperty =
+      typeof projectionVersionProperty === 'function'
+        ? projectionVersionProperty
+        : (p) => p[projectionVersionProperty] as Date | undefined | null;
   }
 
   /**
@@ -68,33 +131,63 @@ export abstract class VersionedEventProcessor<
    * @param event The event from which to build the projection.
    * @param transaction The transaction to use if additional data must be fetched.
    * @param options Options when building the projection.
+   * @returns The projection, and possibly a context with additional information for processing.
    */
   protected abstract project(
     event: E,
     transaction: RWT,
     options?: VersionedProjectionOptions<P>,
-  ): Promise<P>;
+  ): Promise<P | ProjectionWithContext<P, any>>;
 
   /**
-   * Checks whether the given projection is strictly more recent than the state, based on their
+   * Validates that the given projection is strictly more recent than the state, based on their
    * {@link VersionedEventProcessor.projectionVersionProperty} values.
    *
    * @param projection The projection to compare.
    * @param state The state to compare, or `undefined` if it does not exist.
-   * @returns `true` if the projection is more recent than the state, `false` otherwise.
+   * @param options Options for the validation.
    */
-  protected isProjectionMoreRecentThanState(
+  protected validateProjectionIsMoreRecentThanState(
     projection: P,
     state: P | undefined,
-  ): boolean {
+    options: Pick<
+      VersionedEventProcessingOptions<RWT, P>,
+      'reprocessEqualVersion'
+    > &
+      Pick<ProjectionContext<P>, 'projectionVersionProperty'> = {},
+  ): void {
     if (!state) {
-      return true;
+      return;
     }
 
-    return (
-      state[this.projectionVersionProperty] <
-      projection[this.projectionVersionProperty]
-    );
+    const projectionVersionProperty =
+      options.projectionVersionProperty ?? this.projectionVersionProperty;
+
+    const projectionVersion = projectionVersionProperty(projection);
+    if (!projectionVersion) {
+      throw new UnsupportedEntityOperationError(
+        this.projectionType,
+        projection,
+        'Could not find version in projection.',
+      );
+    }
+
+    const stateVersion = projectionVersionProperty(state);
+    if (!stateVersion) {
+      return;
+    }
+
+    const isOld = options.reprocessEqualVersion
+      ? stateVersion > projectionVersion
+      : stateVersion >= projectionVersion;
+    if (isOld) {
+      throw new OldEntityVersionError(
+        this.projectionType,
+        state,
+        projectionVersion,
+        stateVersion,
+      );
+    }
   }
 
   /**
@@ -110,50 +203,102 @@ export abstract class VersionedEventProcessor<
   }
 
   /**
+   * Retrieves the projection for the given key.
+   * If the projection does not exist, an {@link EntityNotFoundError} is thrown.
+   *
+   * @param key The partial projection, containing the properties defining the key for the projection / state.
+   * @param options Options for the operation.
+   * @returns The projection, if it exists.
+   */
+  async get(
+    key: Partial<P>,
+    options: ReadOnlyTransactionOption<ROT> = {},
+  ): Promise<P> {
+    return await this.runner.run(
+      { readOnly: true, transaction: options.transaction },
+      async (transaction) => {
+        const state = await transaction.get(this.projectionType, key);
+        if (!state) {
+          throw new EntityNotFoundError(this.projectionType, key);
+        }
+
+        return state;
+      },
+    );
+  }
+
+  /**
    * Processes the given event, building the corresponding state.
+   * This throws an {@link OldEntityVersionError} if the event's version is older than the existing state.
    *
    * @param event The event to process.
    * @param options Options for processing the event.
-   * @returns The projection if the event was processed, `null` if it was skipped because a more recent state already
-   *   exists.
+   * @returns The projection if the event was processed.
    */
   async processEvent(
     event: E,
-    options: TransactionOption<RWT> & {
-      /**
-       * Whether to skip the check that the projection is more recent than the state.
-       * If `true`, the state will be updated unconditionally.
-       */
-      skipVersionCheck?: boolean;
-    } = {},
-  ): Promise<P | null> {
+    options: VersionedEventProcessingOptions<RWT, P> = {},
+  ): Promise<P> {
     return await this.runner.run(
       { transaction: options.transaction },
       async (transaction) => {
-        const stateKey = this.stateKeyForEvent(event);
-        let state = stateKey
-          ? await transaction.get(this.projectionType, stateKey)
-          : undefined;
+        const projectionWithContext = await this.project(event, transaction, {
+          existingState: options.existingState,
+        });
+        const [projectionOrPartial, context] = Array.isArray(
+          projectionWithContext,
+        )
+          ? projectionWithContext
+          : [projectionWithContext, {}];
 
-        const projection = await this.project(event, transaction, { state });
+        const state =
+          options.existingState !== undefined
+            ? (options.existingState ?? undefined)
+            : context.existingState !== undefined
+              ? (context.existingState ?? undefined)
+              : await transaction.get(this.projectionType, projectionOrPartial);
 
-        if (!options.skipVersionCheck) {
-          state ??= await transaction.get(this.projectionType, projection);
+        const projection = context.defaultProjection
+          ? plainToInstance(this.projectionType, {
+              ...(state ?? context.defaultProjection),
+              ...projectionOrPartial,
+            })
+          : projectionOrPartial;
 
-          const isProjectionMoreRecent = this.isProjectionMoreRecentThanState(
-            projection,
-            state,
-          );
-
-          if (!isProjectionMoreRecent) {
-            return null;
-          }
-        }
+        this.validateProjectionIsMoreRecentThanState(projection, state, {
+          reprocessEqualVersion: options.reprocessEqualVersion,
+          projectionVersionProperty: context.projectionVersionProperty,
+        });
 
         await this.updateState(projection, transaction);
 
         return projection;
       },
     );
+  }
+
+  /**
+   * Processes the given event, building the corresponding state.
+   * This is similar to {@link VersionedEventProcessor.processEvent}, but it returns `null` if the event was skipped
+   * because a more recent state already exists.
+   *
+   * @param event The event to process.
+   * @param options Options for processing the event.
+   * @returns The projection if the event was processed, `null` if it was skipped because a more recent state already
+   *   exists.
+   */
+  async processOrSkipEvent(
+    event: E,
+    options: VersionedEventProcessingOptions<RWT, P> = {},
+  ): Promise<P | null> {
+    try {
+      return await this.processEvent(event, options);
+    } catch (error) {
+      if (error instanceof OldEntityVersionError) {
+        return null;
+      }
+
+      throw error;
+    }
   }
 }