+++ src/versioned-entity/manager.ts
@@ -17,6 +17,7 @@ import {
   Transaction,
   TransactionRunner,
   type ReadOnlyStateTransaction,
+  type ReadOnlyTransactionOption,
   type TransactionOption,
 } from '../transaction/index.js';
 import type { KeyOfType } from '../typing/index.js';
@@ -214,8 +215,11 @@ export class VersionedEntityManager<
 
         transaction.validatePastDateOrFail(existingEntity.updatedAt);
 
-        // This could `null` for a soft-deleted entity, or `undefined` if `deletedAt` is not a property of the entity.
-        if (existingEntity.deletedAt == null) {
+        // This could be `null` for a soft-deleted entity.
+        if (
+          !this.hasDeletionTimestampProperty ||
+          existingEntity.deletedAt === null
+        ) {
           throw new EntityAlreadyExistsError(this.projectionType, entity);
         }
       },
@@ -243,7 +247,11 @@ export class VersionedEntityManager<
           options.existingEntity ??
           (await transaction.get(this.projectionType, entity));
 
-        if (!existingEntity || existingEntity.deletedAt != null) {
+        if (
+          !existingEntity ||
+          (this.hasDeletionTimestampProperty &&
+            existingEntity.deletedAt !== null)
+        ) {
           throw new EntityNotFoundError(this.projectionType, entity);
         }
 
@@ -285,7 +293,14 @@ export class VersionedEntityManager<
   protected async makeProcessAndPublishEvent(
     eventName: EventName<E>,
     entity: EventData<E>,
-    options: VersionedEntityOperationOptions<RWT> = {},
+    options: VersionedEntityOperationOptions<RWT> & {
+      /**
+       * The existing state of the entity for which an event is being created.
+       * If provided, this is passed as the `existingState` to the {@link VersionedEventProcessor} to avoid fetching
+       * the entity from the state again.
+       */
+      existingEntity?: EventData<E> | null;
+    } = {},
   ): Promise<E> {
     return await this.runner.run(
       { transaction: options.transaction },
@@ -309,12 +324,25 @@ export class VersionedEntityManager<
 
         await this.processEvent(event, {
           transaction,
-          skipVersionCheck: true,
+          existingState: options.existingEntity,
         });
 
         return event;
       },
     );
+  }
+
+  async get(
+    key: Partial<EventData<E>>,
+    options: ReadOnlyTransactionOption<ROT> = {},
+  ): Promise<EventData<E>> {
+    const entity = await super.get(key, options);
+
+    if (this.hasDeletionTimestampProperty && entity.deletedAt) {
+      throw new EntityNotFoundError(this.projectionType, key);
+    }
+
+    return entity;
   }
 
   /**
@@ -350,6 +378,7 @@ export class VersionedEntityManager<
         const event = await this.makeProcessAndPublishEvent(eventName, entity, {
           transaction,
           publishOptions: options.publishOptions,
+          existingEntity: null,
         });
 
         return event;
@@ -427,6 +456,7 @@ export class VersionedEntityManager<
         const event = await this.makeProcessAndPublishEvent(eventName, entity, {
           transaction,
           publishOptions: options.publishOptions,
+          existingEntity,
         });
 
         return event;
@@ -477,6 +507,7 @@ export class VersionedEntityManager<
         const event = await this.makeProcessAndPublishEvent(eventName, entity, {
           transaction,
           publishOptions: options.publishOptions,
+          existingEntity,
         });
 
         return event;