+++ CHANGELOG.md
@@ -7,11 +7,20 @@ Breaking changes:
 - Upgrade the minimum Node.js version to `20`.
 - Refactor the `Transaction` and `TransactionRunner`, making all transactions "find replace transactions", using the `get`, `set`, and `delete` terminology.
 - Define a unique `TransactionRunner.run` method which supports existing transactions (replacing `runInNewOrExisting`).
+- Replace `skipVersionCheck` option with `existingState` in the `VersionedEventProcessor`.
+- Make `VersionedEventProcessor.processEvent` throw instead of returning `null`, and introduce the `processOrSkipEvent` method with the old behavior.
+- Allow `VersionedEventProcessor.project` to return a context, which can be used to specify a version property to use, replaces the `stateKeyForEvent` logic, and to support merges of partial projections.
 
 Features:
 
 - Introduce the `ReadOnlyStateTransaction`, along with the corresponding `readOnly` option when creating a new transaction using a runner.
 - Support transaction-wide `publishOptions`.
+- Implement `VersionedEventProcessor` and `VersionedEntityManager` `get` method.
+- Support `reprocessEqualVersion` in the `VersionedEventProcessor` when processing events for backfilling use cases.
+
+Fixes:
+
+- Do not check `deletedAt` when the entity does not have a deletion property.
 
 ## v0.27.2 (2025-05-28)